compute pallete on OpenCL device

+++ Cursa4/Utils.h
@@ -13,6 +13,6 @@ class Utils
 public:
 	static Bitmap^ SequentialChangeContrast(Bitmap^ img, double coeff);
 	static Bitmap^ OmpChangeContrast(Bitmap^ img, double coeff);
-	static Bitmap^ CLChangeContrast(cl_device_id* devices, Bitmap^ img, double coeff);
+	static Bitmap^ CLChangeContrast(cl_device_id* devices, Bitmap^ img, float coeff);
 	static int convertToString(const char* fileName, std::string& s);
 };
